Multiplex SSL and non-SSL connections on the same port
As the IANA has ceased assigning separate TCP ports for SSL/TLS-encrypted
transports for services, it is desirable to be able to connect to a service
via either AMQP or AMQPS on a single port. AMQP-1.0 provides for a
STARTTLS-style mechanism to allow for this, however this is not supported
in AMQP-0.10.

This patch allows Qpid to detect whether a connection uses AMQP or AMQPS by
checking for an SSL/TLS Client Hello at the start of the stream. Since the
header for AMQP connections always starts with the bytes 'AMQP', it is
possible to safely detect an SSL/TLS-wrapped stream.

Detection is enabled when the SSL plugin is active and 'port' and 'ssl-port'
are set to the same value.

Signed-off-by: Zane Bitter <zbitter@redhat.com>